pyproject.toml
[project]
name = "rag-with-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.74.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
